+++ android/src/main/java/com/instabug/flutter/InstabugFlutterPlugin.java
@@ -1,15 +1,17 @@
 package com.instabug.flutter;
 
-import android.annotation.SuppressLint;
 import android.app.Activity;
+import android.app.Application;
 import android.content.Context;
 import android.graphics.Bitmap;
+import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import com.instabug.flutter.generated.InstabugFlutterPigeon;
 import com.instabug.flutter.modules.ApmApi;
 import com.instabug.flutter.modules.BugReportingApi;
 import com.instabug.flutter.modules.CrashReportingApi;
@@ -19,6 +21,7 @@ import com.instabug.flutter.modules.InstabugLogApi;
 import com.instabug.flutter.modules.RepliesApi;
 import com.instabug.flutter.modules.SessionReplayApi;
 import com.instabug.flutter.modules.SurveysApi;
+import com.instabug.flutter.util.ThreadManager;
 
 import java.util.concurrent.Callable;
 
@@ -31,9 +34,72 @@ import io.flutter.plugin.common.BinaryMessenger;
 public class InstabugFlutterPlugin implements FlutterPlugin, ActivityAware {
     private static final String TAG = InstabugFlutterPlugin.class.getName();
 
-    @SuppressLint("StaticFieldLeak")
-    private static Activity activity;
+    private Activity flutterActivity;
 
+    public InstabugFlutterPigeon.InstabugFlutterApi instabugFlutterApi;
+
+    private boolean isLastScreenInstabug = false;
+    Application.ActivityLifecycleCallbacks callbacks = new Application.ActivityLifecycleCallbacks() {
+        @Override
+        public void onActivityCreated(@NonNull Activity activity, @Nullable Bundle savedInstanceState) {
+
+        }
+
+        @Override
+        public void onActivityStarted(@NonNull Activity activity) {
+            if (isLastScreenInstabug) {
+                isLastScreenInstabug = false;
+                return;
+            }
+
+            isLastScreenInstabug = activity.getComponentName().getClassName().contains("com.instabug.");
+            if (activity == flutterActivity) {
+                ThreadManager.runOnMainThread(new Runnable() {
+                    @Override
+                    public void run() {
+                        try {
+                            if (instabugApi != null) {
+                                instabugApi.reportLastScreenChange();
+                            }
+                        } catch (Exception e) {
+                            e.printStackTrace();
+                        }
+                    }
+                });
+
+            }
+
+        }
+
+        @Override
+        public void onActivityResumed(@NonNull Activity activity) {
+
+        }
+
+
+        @Override
+        public void onActivityPaused(@NonNull Activity activity) {
+
+        }
+
+        @Override
+        public void onActivityStopped(@NonNull Activity activity) {
+
+
+        }
+
+        @Override
+        public void onActivitySaveInstanceState(@NonNull Activity activity, @NonNull Bundle outState) {
+
+        }
+
+        @Override
+        public void onActivityDestroyed(@NonNull Activity activity) {
+
+        }
+    };
+
+    private InstabugApi instabugApi;
 
     @Override
     public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
@@ -42,30 +108,40 @@ public class InstabugFlutterPlugin implements FlutterPlugin, ActivityAware {
 
     @Override
     public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
-        activity = null;
+        flutterActivity = null;
     }
 
     @Override
     public void onAttachedToActivity(@NonNull ActivityPluginBinding binding) {
-        activity = binding.getActivity();
+        flutterActivity = binding.getActivity();
+        initOnActivityResultListener();
     }
+
+    private void initOnActivityResultListener() {
+        Application app = flutterActivity.getApplication();
+        app.unregisterActivityLifecycleCallbacks(callbacks);
+        app.registerActivityLifecycleCallbacks(callbacks);
+    }
+
 
     @Override
     public void onDetachedFromActivityForConfigChanges() {
-        activity = null;
+        flutterActivity = null;
+
     }
 
     @Override
     public void onReattachedToActivityForConfigChanges(@NonNull ActivityPluginBinding binding) {
-        activity = binding.getActivity();
+        flutterActivity = binding.getActivity();
+        initOnActivityResultListener();
     }
 
     @Override
     public void onDetachedFromActivity() {
-        activity = null;
+        flutterActivity = null;
     }
 
-    private static void register(Context context, BinaryMessenger messenger, FlutterRenderer renderer) {
+    private void register(Context context, BinaryMessenger messenger, FlutterRenderer renderer) {
         final Callable<Bitmap> screenshotProvider = new Callable<Bitmap>() {
             @Override
             public Bitmap call() {
@@ -77,17 +153,19 @@ public class InstabugFlutterPlugin implements FlutterPlugin, ActivityAware {
         BugReportingApi.init(messenger);
         CrashReportingApi.init(messenger);
         FeatureRequestsApi.init(messenger);
-        InstabugApi.init(messenger, context, screenshotProvider);
+        instabugApi = InstabugApi.init(messenger, context, screenshotProvider);
         InstabugLogApi.init(messenger);
         RepliesApi.init(messenger);
         SessionReplayApi.init(messenger);
         SurveysApi.init(messenger);
+        instabugFlutterApi = new InstabugFlutterPigeon.InstabugFlutterApi(messenger);
+
     }
 
     @Nullable
-    private static Bitmap takeScreenshot(FlutterRenderer renderer) {
+    private Bitmap takeScreenshot(FlutterRenderer renderer) {
         try {
-            final View view = activity.getWindow().getDecorView().getRootView();
+            final View view = flutterActivity.getWindow().getDecorView().getRootView();
 
             view.setDrawingCacheEnabled(true);
             final Bitmap bitmap = renderer.getBitmap();