+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased](https://github.com/Instabug/Instabug-Flutter/compare/v14.3.0...dev)
+
+### Added
+
+- Add support for xCode 16. ([#574](https://github.com/Instabug/Instabug-Flutter/pull/574))
+
 ## [14.3.0](https://github.com/Instabug/Instabug-Flutter/compare/v14.1.0...14.3.0) (April 21, 2025)
 
 ### Added