+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased](https://github.com/Instabug/Instabug-Flutter/compare/v14.3.0...dev)
+
+### Added
+
+- Add support for hybrid apps. ([#561](https://github.com/Instabug/Instabug-Flutter/pull/561))
+
+
 ## [14.3.0](https://github.com/Instabug/Instabug-Flutter/compare/v14.1.0...14.3.0) (April 21, 2025)
 
 ### Added